refactor: export progress bar to dedicated component file
